Copyright:
After further investigation it appears that Ben Fennema was the CVS
committer of chkudf and not its author. This could explain why chkudf
was not chosen for migration to udftools at the time that other userland
tools were spun off.

The source code was fully formed in the initial commit in November 1999,
and did not change very much (certainly without any detailed commit logs)
thereafter. As part of the initial commit to CVS the Makefile at a higher
level was changed to add code to print "Making Rob Simm's chkudf."

Based on that and the CREDITS file higher up, reassign original copyright to
Rob Simms for all files except do_pkt.c (and HP-copyrighted files).
When do_pkt.c was created in September 2000 neither it nor the CVS commitlog
contained any information that could be used to determine copyright.